+++ CHANGELOG.md
@@ -5,10 +5,13 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Changed
 * Rollback `semantic-ui-ember` addon to v0.9.3.
+* Renamed property for input in `flexberry-file`. Deprecated `class`, new `inputClass`.
 
 ### Fixed
 * `flexberry-checkbox` component:
     * Now with `semantic-ui-ember` v0.9.3, if `flexberry-checkbox` was unchecked it doesn't have `checked` class.
+* `edit-form` controller: add omponentProperties for `flexberry-file`.
+
 * Now `flexberry-file`'s download button is disabled after deleting selected file and saving model.
 * Filters for `flexberry-objectlistview` are working now (tere was a code that was accidentally deleted during merges).
 * Transition to `edit-form` route after saving new details.